requirements.txt
groq
fastapi
python-dotenv
transformers
sentencepiece
protobuf
torch
llama-index
llama-index-embeddings-openai

